- replacing virtual class fcontainer by templates - rewriting mesh::load_bnd to handle ASA comments - adding tool to convert mesh to dipole file (om_mesh_to_dip) - renaming vect3array class to Vect3array for consistency with Vect3 class

+++ tools/make_nerve.cpp
@@ -110,7 +110,7 @@ decoupe(dt,iz2,&niz2,&diz2);
 
   float ang[c];
 
-  int max= floor(nteta*(L/dt+2*c)*4);
+  int max = (int) floor(nteta*(L/dt+2*c)*4);
     Vect3* P = new Vect3[max];
     Triangle* T = new Triangle[2*max];
